requirements.txt
Django==4.2.2
django-cors-headers==4.2
djangorestframework==3.14
